Move eucameta.py to prevent sosreport spurious output
When 'sosreport -l' was run on a system where eucalyptus-sosreport-plugins
was installed, it would print extra the warning "plugin eucameta does not install".

This was because the eucameta.py really only existed to maintain the version
string for eucalyptus-sosreport-plugins, and did not actually perform sosreport-y
work.  We will correct that by moving the version string information into VERSION.py
so it won't impede sosreport in any way.